+++ src/bedrock_server_manager/web/app.py
@@ -25,7 +25,9 @@ logger = logging.getLogger(__name__)
 
 
 def run_web_server(
-    host: Optional[Union[str, List[str]]] = None, debug: bool = False
+    host: Optional[Union[str, List[str]]] = None,
+    debug: bool = False,
+    threads: Optional[int] = None,
 ) -> None:
     """
     Configures and starts the Uvicorn web server to serve the FastAPI application.
@@ -150,13 +152,20 @@ def run_web_server(
     else:
         threads_setting_key = "web.threads"
         try:
-            workers_val = int(settings.get(threads_setting_key, 4))
-            if workers_val > 0:
-                workers = workers_val
+            if threads is None:
+                workers_val = int(settings.get(threads_setting_key, 4))
+                if workers_val > 0:
+                    workers = workers_val
+                else:
+                    logger.warning(
+                        f"Invalid '{threads_setting_key}' ({workers_val}). Using default: {workers}."
+                    )
             else:
-                logger.warning(
-                    f"Invalid '{threads_setting_key}' ({workers_val}). Using default: {workers}."
-                )
+                workers = int(threads)
+                if workers <= 0:
+                    logger.warning(
+                        f"Invalid 'threads' passed ({workers}). Using default: {workers}."
+                    )
         except (ValueError, TypeError):
             logger.warning(
                 f"Invalid format for '{threads_setting_key}'. Using default: {workers}."